requirements.txt
PyMuPDF>=1.24.0
google-generativeai>=0.8.0
requests>=2.31.0
pytest>=8.0.0
